@@ -24,7 +24,7 @@
     "@codemirror/search": "^6.0.0",
     "@codemirror/state": "^6.0.0",
     "@codemirror/view": "^6.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.1",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "chart.js": "^4.0.0",
     "chartjs-adapter-luxon": "^1.2.0",
     "chartjs-plugin-zoom": "^2.0.1",
@@ -32,7 +32,7 @@
     "luxon": "^3.4.4",
     "pocketbase": "^0.25.1",
     "sass": "^1.45.0",
-    "svelte": "^4.0.0",
+    "svelte": "^5.0.0",
     "svelte-flatpickr": "^3.3.3",
     "svelte-spa-router": "^4.0.0",
     "vite": "^5.0.11"
