@@ -19,11 +19,11 @@
 
     let oldLocation = undefined;
 
-    let showAppSidebar = false;
+    let showAppSidebar = $state(false);
 
-    // theme
+    // theme - using runes for local state
     const THEME_KEY = "pb_ui_theme";
-    let isDark = false;
+    let isDark = $state(false);
 
     function applyTheme(dark) {
         isDark = !!dark;
@@ -56,11 +56,14 @@
         window.localStorage?.setItem(THEME_KEY, isDark ? "dark" : "light");
     }
 
-    let isTinyMCEPreloaded = false;
+    let isTinyMCEPreloaded = $state(false);
 
-    $: if ($superuser?.id) {
-        loadSettings();
-    }
+    // Using Svelte 5 $effect for reactive side effects
+    $effect(() => {
+        if ($superuser?.id) {
+            loadSettings();
+        }
+    });
 
     function handleRouteLoading(e) {
         if (e?.detail?.location === oldLocation) {
@@ -142,7 +145,7 @@
                     use:active={{ path: "/collections/?.*", className: "current-route" }}
                     use:tooltip={{ text: "Collections", position: "right" }}
                 >
-                    <i class="ri-database-2-line" />
+                    <i class="ri-database-2-line"></i>
                 </a>
                 <a
                     href="/logs"
@@ -152,7 +155,7 @@
                     use:active={{ path: "/logs/?.*", className: "current-route" }}
                     use:tooltip={{ text: "Logs", position: "right" }}
                 >
-                    <i class="ri-line-chart-line" />
+                    <i class="ri-line-chart-line"></i>
                 </a>
                 <a
                     href="/settings"
@@ -162,7 +165,7 @@
                     use:active={{ path: "/settings/?.*", className: "current-route" }}
                     use:tooltip={{ text: "Settings", position: "right" }}
                 >
-                    <i class="ri-tools-line" />
+                    <i class="ri-tools-line"></i>
                 </a>
             </nav>
 
@@ -185,11 +188,11 @@
                         role="menuitem"
                         use:link
                     >
-                        <i class="ri-shield-user-line" aria-hidden="true" />
+                        <i class="ri-shield-user-line" aria-hidden="true"></i>
                         <span class="txt">Manage superusers</span>
                     </a>
-                    <button type="button" class="dropdown-item closable" role="menuitem" on:click={logout}>
-                        <i class="ri-logout-circle-line" aria-hidden="true" />
+                    <button type="button" class="dropdown-item closable" role="menuitem" onclick={logout}>
+                        <i class="ri-logout-circle-line" aria-hidden="true"></i>
                         <span class="txt">Logout</span>
                     </button>
                 </Toggler>
@@ -200,7 +203,7 @@
                 class="btn btn-transparent btn-sm m-t-10"
                 aria-pressed={isDark}
                 title="Toggle dark theme"
-                on:click={toggleTheme}
+                onclick={toggleTheme}
             >
                 {#if isDark}
                     <i class="ri-sun-line" aria-hidden="true"></i>
