@@ -17,7 +17,7 @@
 
         {#if $superuser?.id}
             <a href={import.meta.env.PB_DOCS_URL} target="_blank" rel="noopener noreferrer">
-                <i class="ri-book-open-line txt-sm" />
+                <i class="ri-book-open-line txt-sm"></i>
                 <span class="txt">Docs</span>
             </a>
             <span class="delimiter">|</span>
